pyproject.toml
[project]
name = "catalogaiapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.1",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "pymongo>=4.16.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
